backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-decouple
Pillow
openpyxl
